Removed redundant folders and files
- no more composables
- testStores>test
